webserver/requirements.txt
setuptools==57.5.0
chardet==3.0.4
Flask==1.1.2
Flask-Bcrypt==0.7.1
flask-restx==0.5.1
flask-expects-json==1.7.0
funcy==1.14
importlib-metadata==1.6.0
itsdangerous==1.1.0
Jinja2==2.11.1
jmespath==0.9.5
jsonschema==3.2.0
MarkupSafe==1.1.1
pip-tools==4.5.1
requests==2.23.0
retry==0.9.2
urllib3==1.25.8
Werkzeug==0.16.1
wsgi-request-logger==0.4.6
Authlib==0.14.3
Flask-JWT-Extended==4.0.2
Flask-SQLAlchemy==2.4.4
SQLAlchemy==1.4.3
Flask-Migrate==2.6.0
Flask-Login==0.5.0
stringcase==1.2.0
Flask-Cors==3.0.10
gunicorn==20.1.0
pymongo==4.1.1
ed25519==1.5
PyNaCl==1.5.0
pika==1.3.0
psycopg2==2.9.3
pytz==2022.2.1
python-logging-loki==0.3.1
